Parse geolocation data correctly
Geolocation is represented as a JSON object with either a place-id
field, or latitude/longitude fields. We were previously trying to put
this whole object in to the database, which did not work.
This commit splits out the fields of the geolocation object.

Removed the `menu_ui` dependency, which is now part of core.