Get maximum number of events from libcal
When using calendar ids there was no limit specified, so it was using
the default of 20. Increased it to the maximum possible of 100. The
libcal event_search API does not support pagination.

Reversed the logic so the `-1` functionality works.

Refactored the code so it's more concise. There's no need to guard
`explode`. It will still create an array with a single entry if the
delimiter is not present.